change freeformat converter to findent

+++ bin/upgrade_fortran.py
@@ -14,6 +14,12 @@
 # - go to the source folder
 # - run upgrade_fortran.py *.f
 # - powergrep include extensions ".inc" => ".inc90"
+#
+# new interface:
+# upgrade_fortran check *.f
+# upgrade_fortran freeformat *.f
+# upgrade_fortran split *.f90
+
 
 import sys
 import os
@@ -25,28 +31,124 @@ import glob
 sys.path.append(r'C:\msys64\mingw64\bin')
 f90ppr_exe = r"F:\f90ppr\moware\f90ppr"
 f90split_exe = r"F:\f90ppr\moware\f90split"
+findent_exe = r"F:\findent-3.1.6\findent"
 
 
-def main(f77name):
+def check_one(f77name):
+    """ performs some preliminary checks in the source files
+    """
+    # checks line length (f90ppr may truncate long lines of comments)
+    maxlen = 0
+    maxno = 0
+    warns = []
+    with open(f77name, 'rb') as infile:
+        for i,l in enumerate(infile.readlines()):
+            # checks line length
+            clen = len(l)
+            if clen>132:
+                warns.append(f'{os.path.basename(f77name)}:{i+1} line exceeds 132 columns!')
+            if clen>maxlen:
+                maxlen = clen
+                maxno = i+1
+            # checks some bad patterns
+            if b'dowhile' in l.lower():
+                warns.append(f'{os.path.basename(f77name)}:{i+1} replace "dowhile" by "do while"!\n\t'+l.decode().strip())
+            # if b'type' in l.lower():
+            # faire une regex plus subtile! (supprimer commentaires, chaines, variables "typeel")
+            #     warns.append(f'{os.path.basename(f77name)}:{i+1} "type" is a reserved keyword in f90!\n\t'+l.decode().strip())
 
-    # convert file to absolute path
-    f77name = os.path.abspath(f77name)
+    #print(f'longest line ({maxlen} chars) at line {maxno}')
+    for w in warns:
+        print(w)
 
-    # checks that the file exists
-    if not os.path.isfile(f77name):
-        raise Exception(f'{f77name} not found!')
 
-    # CONVERT .f fixed-format file to .f90 free-format -------------------------
+def pretty_one(f90name, keepf=False):
+    """ process .f90 through f90ppr
+    """
 
-    # only processes .f
-    base, ext = os.path.splitext(f77name)
-    if ext != '.f':
-        print(f'ignoring {f77name}')
-        return
+    base, ext = os.path.splitext(f90name)
     f90name = base+'.f90'
 
+    # rename orig file to .f90.bak
+    bakfile = f90name+'.bak'
+    if os.path.isfile(bakfile):
+        os.remove(bakfile)
+    os.rename(f90name, bakfile)
+
     # processes the file with "f90ppr"
-    print(f'f90ppr {f77name} => {f90name}')
+    print(f'f90ppr {os.path.basename(f90name)}')
+    with open(f90name, 'wb') as f90file:
+        cmd = [f90ppr_exe]
+        p = subprocess.Popen(cmd, stdin=subprocess.PIPE, stdout=f90file)
+        # maximum  line length (2-132)
+        p.stdin.write(b'$define FPPR_MAX_LINE 120\n') # 132 produces errors (missing '&' at the EOL!)
+        # keywords case: FPPR_LEAVE, FPPR_UPPER, FPPR_LOWER
+        p.stdin.write(b'$define FPPR_KWD_CASE FPPR_LOWER\n')
+        # variables case: FPPR_LEAVE, FPPR_UPPER, FPPR_LOWER
+        p.stdin.write(b'$define FPPR_USR_CASE FPPR_LEAVE\n')
+        # indentation (0-60)
+        p.stdin.write(b'$define FPPR_STP_INDENT 4\n')
+        # input format:  0=free format
+        p.stdin.write(b'$define FPPR_FXD_IN 0\n')
+        # output format:  0=free format
+        p.stdin.write(b'$define FPPR_FXD_OUT 0\n')
+        with open(bakfile, 'rb') as infile:
+            for l in infile.readlines():
+                p.stdin.write(l)
+        p.stdin.close()
+        retcode = p.wait()
+        if(retcode != 0):
+            print(f'f90ppr ERROR: retcode = {retcode}')
+
+    # remove/rename the .f90.bak file
+    if not keepf:
+        print(f'rm {bakfile}')
+        os.remove(bakfile)
+
+
+def freeformat_one(f77name, keepf=False):
+    """ CONVERT .f fixed-format file to .f90 free-format with findent
+    """
+
+    base, ext = os.path.splitext(f77name)
+    f90name = base+'.f90'
+
+    # do checks
+    check_one(f77name)
+
+    # processes the file with "f90ppr"
+    print(f'findent {os.path.basename(f77name)} => {os.path.basename(f90name)}')
+    with open(f90name, 'wb') as f90file:
+        with open(f77name, 'rb') as infile:
+            cmd = [findent_exe, '-i4', '-ofree']
+            p = subprocess.Popen(cmd, stdin=infile, stdout=f90file)
+            retcode = p.wait()
+            if(retcode != 0):
+                print(f'findent ERROR: retcode = {retcode}')
+
+    # remove/rename the .f file
+    if not keepf:
+        print(f'rm {f77name}')
+        os.remove(f77name)
+    else:
+        bak = f77name+'.bak'
+        os.rename(f77name, bak)
+
+def freeformat_f90ppr_one(f77name, keepf=False):
+    """ CONVERT .f fixed-format file to .f90 free-format with f90ppr
+    this routine do not read comments after line 72 in free format
+    if these comments are placed after a command
+    (comments spanning a whole line are OK)
+    """
+
+    base, ext = os.path.splitext(f77name)
+    f90name = base+'.f90'
+
+    # do checks
+    check_one(f77name)
+
+    # processes the file with "f90ppr"
+    print(f'f90ppr {os.path.basename(f77name)} => {os.path.basename(f90name)}')
     with open(f90name, 'wb') as f90file:
         cmd = [f90ppr_exe]
         p = subprocess.Popen(cmd, stdin=subprocess.PIPE, stdout=f90file)
@@ -70,12 +172,19 @@ def main(f77name):
         if(retcode != 0):
             print(f'f90ppr ERROR: retcode = {retcode}')
 
-    # remove the .f file
-    print(f'rm {f77name}')
-    os.remove(f77name)
+    # remove/rename the .f file
+    if not keepf:
+        print(f'rm {f77name}')
+        os.remove(f77name)
+    else:
+        bak = f77name+'.bak'
+        os.rename(f77name, bak)
 
-    # -- SPLIT THE FILE INTO SUBROUTINES ---------------------------------------
 
+def split_one(f90name, keepf=False):
+    """ SPLIT .f90 files, 1 file per subroutine
+    """
+ 
     # runs f90split from another folder
     # otherwise, routines will have a bad name if the initial file already has
     # the name of a subroutine
@@ -101,7 +210,7 @@ def main(f77name):
     p.stdin.close()
     retcode = p.wait()
     if(retcode != 0):
-        print(f'f90split ERROR: retcode = {retcode}')
+        raise Exception(f'f90split ERROR: retcode = {retcode}')
 
     # check that we do not have any "main000" files
     for f in os.listdir('.'):
@@ -111,8 +220,12 @@ def main(f77name):
             raise Exception('something went wrong!')
 
     # remove the .f90 file
-    print(f'rm {f90name}')
-    os.remove(f90name)
+    if not keepf:
+        print(f'rm {f90name}')
+        os.remove(f90name)
+    else:
+        bak = f90name+'.bak'
+        os.rename(f90name, bak)  
 
     # copy subroutines to initial folder
     f90folder = os.path.dirname(f90name)
@@ -130,8 +243,73 @@ def main(f77name):
     os.chdir(origdir)
 
 
+def iterate(files, exts=['.f', '.for', '.f90','.inc']):
+    """ iterates over the files with given extension and 
+    performs some checks. 
+    Yields valid absolute filenames one by one.
+    """
+    for f in files:
+        for gf in glob.glob(f):
+            # convert file to absolute path
+            gf = os.path.abspath(gf)
+            # check extension
+            base, ext = os.path.splitext(gf)
+            if not ext in exts:
+                print(f'ignoring {gf}')
+                continue
+            # check whether file exists
+            if not os.path.isfile(gf):
+                raise Exception(f'{gf} not found!')
+            yield gf
+
+
+def check(files):
+    for f in iterate(files, exts=['.f', '.for','.inc']):
+        print(f'checking file {f}')
+        check_one(f)
+
+def split(files, keep):
+    for f in iterate(files, exts=['.f90']):
+        print(f'splitting file {f}')
+        split_one(f, keep)
+
+def freeformat(files, keep):
+    for f in iterate(files, exts=['.f', '.for', '.f90']):
+        print(f'converting file {f} to free format')
+        freeformat_one(f, keep)
+
+def pretty(files, keep):
+    for f in iterate(files, exts=['.f90']):
+        print(f'f90ppr {f}')
+        pretty_one(f, keep)
+
+
 if __name__ == "__main__":
-    
+
+
+    # parse cmd-line arguments
+    import argparse
+    parser = argparse.ArgumentParser(description='Upgrade old FORTRAN.')
+    parser.add_argument("--keep", help="keep original files", action="store_true")
+    # parser.add_argument("--include", help="include pattern", default='')
+    # parser.add_argument("--exclude", help="exclude pattern", default='')
+    parser.add_argument('command', help='command', choices=[
+                        'check', 'split', 'freeformat', 'pretty'])
+    parser.add_argument('files', nargs='+', help='fortran files')
+    args = parser.parse_args()
+    print (args)
+
+    if args.command == 'check':
+        check(args.files)
+    elif args.command == 'split':
+        split(args.files, args.keep)
+    elif args.command == 'freeformat':
+        freeformat(args.files, args.keep)
+    elif args.command == 'pretty':
+        pretty(args.files, args.keep)
+    else:
+        raise Exception("Unknown arg: {}".format(args.command))
+
+
     # main(sys.argv[1])
-    for f in glob.glob(sys.argv[1]):
-        main(f)
+
